Refactored StatUploader into __init__.py in stat folder.
